+++ wp-appkit/lib/themes/themes.php
@@ -18,12 +18,12 @@ class WpakThemes {
 
 		'q-ios' => array(
 			'name' => 'Q for iOS',
-			'version' => '1.0.5',
+			'version' => '1.0.6',
 		),
 
 		'q-android' => array(
 			'name' => 'Q for Android',
-			'version' => '1.0.5',
+			'version' => '1.0.6',
 		),
 
 	);